# backend/openapi.yaml
openapi: 3.0.3
info:
  title: DockerFleet Manager API
  description: |
    Backend API for DockerFleet Manager. Most endpoints require Bearer JWT authentication.
    Obtain a token via POST /auth/login or POST /auth/register.
    For API scanners: fetch this spec from GET {baseUrl}/api-docs/openapi.json (e.g. http://localhost:5000/api-docs/openapi.json)
    and set the server base URL to your app host (e.g. http://localhost:5000). Use the token from login in the Authorization header.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API base path; use full base URL in scanner (e.g. http://localhost:5000)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Invalid credentials"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
...
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        refreshToken: { type: string }
        user:
          type: object
          properties:
            id: { type: integer }
            email: { type: string }
            role: { type: string, enum: [admin, user] }
...
    ExecuteCommand:
      type: object
      required: [command]
      properties:
        command: { type: string }
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/register:
    post:
      tags: [Auth]
      summary: Register (first user becomes admin)
      security: []
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        "400":
          description: User already exists or validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user (requires Bearer)
      responses:
        "200":
          description: Current user
        "401":
          description: Unauthorized

  # ---------- Users ----------
...
  /servers/{serverId}/containers/{containerId}/execute:
    post:
      tags: [Containers]
      summary: Execute command in container
      parameters:
        - name: serverId
          in: path
          required: true
          schema: { type: string }
        - name: containerId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExecuteCommand' }
      responses:
        "200":
          description: Command output
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Not found

  # ---------- Images ----------
...